user configured `py-shell-name' is not honored by `py-guess-default', lp:920079, fixed
new command `py-execute-region-default' forces the
systems default Python interpreter to execute, ignores
shebang

related functions redesigned, `async' argument dropped
`py-execute-region' now reads:

(defun py-execute-region (start end &optional shell dedicated)
  "Send the region to a Python interpreter.

When called with \\[univeral-argument], execution through `default-value' of `py-shell-name' is forced.
When called with \\[univeral-argument] followed by a number different from 4 and 1, user is prompted to specify a shell. This might be the name of a system-wide shell or include the path to a virtual environment.

When called from a programm, it accepts a string specifying a shell which will be forced upon execute as argument. "
[ ... ]